restructured directory for package distribution
