+++ omrmodules/normalization/sample_photo.jpg
