+++ omrmodules/semantics/__init__.py
@@ -1,0 +1,1 @@
+print('imported semantics')
