+++ omrmodules/semantics/samples/measuredetections.dict
