+++ omrmodules/semantics/samples/sample_image.png
