+++ omrmodules/visionutils/utils.py
